+++ deploy.sh
@@ -1,8 +1,9 @@
-# git status
-# git pull
+git status
+git checkout production
+git pull
 
 # Install all or some packages
-# php ../composer.phar install
+php composer update
 
 # Add changes from entities to database
-# php doctrine orm:schema-tool:update --force
+php doctrine orm:schema-tool:update --force