+++ src/localisation.php
@@ -1,12 +1,13 @@
 <?php
-if (!$settings['settings']['locale']['use_domain'] 
-        && $settings['settings']['locale']['process'] === 'session') {
+if (($settings['settings']['locale']['process'] & \App\Utility\LanguageUtility::DOMAIN_DISABLED) == \App\Utility\LanguageUtility::DOMAIN_DISABLED 
+        && ($settings['settings']['locale']['process'] & \App\Utility\LanguageUtility::LOCALE_SESSION) == \App\Utility\LanguageUtility::LOCALE_SESSION) {
     if (isset($_COOKIE['current_locale'])) {
         $settings['settings']['locale']['code'] = $_COOKIE['current_locale'];
     } else {
         setcookie('current_locale', $settings['settings']['locale']['code'], 0, '/');
     }
-} elseif ($settings['settings']['locale']['use_domain']) {
+} elseif (($settings['settings']['locale']['process'] & \App\Utility\LanguageUtility::DOMAIN_ENABLED) == \App\Utility\LanguageUtility::DOMAIN_ENABLED 
+        && ($settings['settings']['locale']['process'] & \App\Utility\LanguageUtility::LOCALE_URL) == \App\Utility\LanguageUtility::LOCALE_URL) {
     $uri = $_SERVER['SERVER_NAME'];
 
     if (array_search($uri, $settings['settings']['locale']['active']) === FALSE) {
@@ -14,7 +15,8 @@ if (!$settings['settings']['locale']['use_domain']
     } else {
         $settings['settings']['locale']['code'] = array_search($uri, $settings['settings']['locale']['active']);
     }
-} else {
+} elseif (($settings['settings']['locale']['process'] & \App\Utility\LanguageUtility::DOMAIN_DISABLED) == \App\Utility\LanguageUtility::DOMAIN_DISABLED 
+        && ($settings['settings']['locale']['process'] & \App\Utility\LanguageUtility::LOCALE_URL) == \App\Utility\LanguageUtility::LOCALE_URL) {
     // if domain points to public directory
     if ($settings['settings']['public_path'] == '/') {
         $uri = substr($_SERVER['REQUEST_URI'], 1, 6);
@@ -39,4 +41,6 @@ if (!$settings['settings']['locale']['use_domain']
             $settings['settings']['locale']['code'] = 'en-US';
             break;
     }
+} else {
+    die('Locale process setting not allowed');
 }